Add an option to treat NULL keys as equal in joins (GPUDB/gqe!42)
- Add an argument `compare_nulls` to the logical join relation, the physical join relation and the join task.

By default, SQL treats NULLs as not equal in joins, but as equal in set operations (like INTERSECT). Before this MR, NULL keys are always compared not equal for joins in GQE. Since joins can also be used to implement INTERSECT, this MR adds flexibility to specify whether the NULL keys should compare equal or not.