Extended expression support (GPUDB/gqe!44)
This MR enables expressions that contain string literals, as well as expressions that evaluate to non fixed width result columns.

tl;dr we split a given input AST into sub-trees that can be evaluated efficiently using `cudf::ast` and fallback to a custom evaluation strategy for AST nodes that are not supported by `cudf::ast`.