Add in-memory storage infrastructure (Devtech-Compute/gqe!85)
This is the first MR of a series that will add in-memory storage to GQE.

The goal of in-memory storage is to run queries on in-memory tables. This MR adds the fundamental infrastructure to support multiple table kinds. Specifically, it adds a unified table resource `table` that provides a table management interface, and a unified table access method `readable_view` that provides an interface to read data from a table.

The initial implementation supports Parquet files.