Gather all Parquet storage classes in the Parquet storage file (Devtech-Compute/gqe!116)
This MR refactors and cleans up the Parquet reader and writer.

Previously, these were defined alongside the read and write base classes. However, the base classes are also used by the in-memory storage, and in future more storage kinds will be added.

Thus, all classes related to Parquet storage are now in the same place.