Remove the deprecated `catalog::register_table()` overload (Devtech-Compute/gqe!117)
The new `catalog::register_table()` function introduced with the in-memory
storage infrastructure is general and can handle all current use-cases.

The old overloaded function is no longer needed and is effectively deprecated.
However, it was still in use by some benchmarks.

This MR refactors all code relying on the deprecated function to use the new
function, and removes the deprecated code in the `catalog`.

This MR closes issue https://gitlab-master.nvidia.com/Devtech-Compute/gqe/-/issues/52.