Add in-memory support for Substrait TPC queries (Devtech-Compute/gqe!114)
With this MR, TPC-{DS,H} queries can execute using in-memory tables.

When running a benchmark, the user can now select a storage kind on the command line. If an in-memory table kind is selected, GQE will copy all Parquet input files into corresponding in-memory tables. Then, GQE runs the benchmark on the these in-memory tables.

This MR also includes fixes the TPC-DS schema to match the Parquet files generated by @haog.