Add Docker container for aarch64 (Devtech-Compute/gqe!128)
# Overview

The immediate goal of this MR are G+H measurements. For these measurements, it's necessary to have a consistent build environment for GQE on both the `x86_64` and the `aarch64` platforms.

GQE is typically built and run in a Docker container. The reason is that GQE should be run with CUDA per-thread default stream (PTDS) enabled for the best performance. We build libcudf from source to enable the PTDS feature, and distribute the complete environment for GQE in the Docker container.

# Changes

This MR adds `aarch64` support in the Dockerfile. To make maintaining the Docker container easier, we use the RAPIDS devcontainer as a base image instead of the CUDA Docker container.

The dependencies for libcudf and GQE are installed only using the `apt` repository, instead of mixing `apt` and `conda-forge`. This avoids package conflicts, and allows us to use the configuration tested upstream by the RAPIDS team for libcudf.

Mamba is available in the Docker container if needed for development.

The size of the Docker containers are reduced by about `2x`, from ~10 GB to ~4.5 GB. This is a result of 

# Testing

This container has been tested on G+H (`aarch64`) by building GQE, running `make test`, and running TPC-H queries on SF100. Also tested is building GQE on `x86_64` and running `make test`.

# Container Availability

The container images are available as `gitlab-master.nvidia.com:5005/devtech-compute/gqe:23.08-x86_64` and `gitlab-master.nvidia.com:5005/devtech-compute/gqe:23.08-aarch64`.

# Future Work

This MR does not address the following issues, which are left to future MRs:

 - `-lineinfo` is disabled to work around a [CUDA compiler bug](https://nvidia.slack.com/archives/CBB1XVATD/p1692710320243829?thread_ts=1673635966.417389&cid=CBB1XVATD) for `sm_90`. The bug is fixed in CUDA 12.2. Thus, `-lineinfo` should be re-enabled when upgrading.
 - Protobuf 3.12 caused a GQE build failure. This MR removes the affected lines of code. These could be reintroduced later on, as they check an error condition.
 - [Libcudf doesn’t build with Arrow S3 enabled](https://gitlab-master.nvidia.com/Devtech-Compute/gqe/-/merge_requests/128#note_16954343).
 - A Docker manifest could unify the x86_64 and aarch64 containers. The standard way of building an multi-arch container is to use `docker buildx build`. This requires some extra configuration on the build servers to either use native `x86_64` and `aarch64` servers, or use QEMU `aarch64` emulation on an `x86_64` build server. `docker buildx` was [prototyped on Blossom](https://confluence.nvidia.com/display/BLOS/BuildX+Support), but it's not clear from the documentation whether it's a supported tool.