Use floating point types instead of decimal types consistently (Devtech-Compute/gqe!145)
GQE currently does not support decimal types properly, so it *sometimes* uses floating-point types instead. For example, decimal literals are parsed as floating points in the Substrait consumer. However, the inconsistencies between floating-points and decimals are sometimes problematic. For example, cuDF's AST module cannot evaluate binary expressions when one side is of a decimal type and the other side is of a floating-point type.

This MR makes GQE always uses floating-point type. In the future, we should properly support decimal types and expose whether to use floating-point or decimal types as a tuning knob to the user.

Specifically, this MR implements the following.
- Cast expressions to decimal types are parsed as cast expressions to `FLOAT64` in the Substrait consumer.
- Support cast expressions in the join condition. Integral types are casted to `INT64`, and all floating-point and decimal types are casted to `FLOAT64`.
- Enable TPC-H Q11, Q17 and Q20 in the CI.

Close #73