Refactored plan comparison utilities from optimizer implementation draft (Devtech-Compute/gqe!151)
To test whether the optimizer produces the expected plans, we need to be able to compare two logical plans and return `true` if they look exactly the same.

This code was partially part of the logical optimizer implementation draft. However, as the optimizer is being developed and tested, the code grew. Thus, I've decided to extract the code for plan comparison from the optimizer branch and tested it separately. The result is this MR.