Optimize performance for Hive partitioned datasets (Devtech-Compute/gqe!166)
Before this MR, GQE will construct the partition key column in a Hive partitioned dataset by decompressing and decoding the column in the Parquet file. However, this is not necessary because we can get the values of the partition key column from the directory names.

This MR
- Constructs the partition key column in memory from the directory names instead of loading from Parquet files.
- Refactors the partial filter parsing in the Parquet read task.

In order to construct the partition key column, we need to know how many rows are in each Parquet file. For GQE's customized Parquet reader, we can get this information for free, since we need to decode the file footer anyway as part of the Parquet reading process. However, cuDF's Parquet reader does not return the number of rows in each file, so we need to determine that out-of-band, which brings extra overhead.

Close https://gitlab-master.nvidia.com/Devtech-Compute/gqe/-/issues/26.