Integer overflow fix in groupby implementation (Devtech-Compute/gqe!186)
This MR fixes the integer overflow bug in the groupby implementation.

The `num_keys` is of `cudf::size_type` which is `int32_t`, and the operation `num_keys * 2` overflows when `num_keys` is roughly greater than 1 Million.

This leads to the static set using `int64_t` as the extent type. This has less than 1-2% regression measured on SF100 on G+H.