Add MLIR as CMake dependency along with integration tests (Devtech-Compute/gqe!195)
[MLIR](https://mlir.llvm.org) is a framework for building compilers. In GQE, we will use it as a basis for a just-in-time query compiler. MLIR supports us in this endeavor by providing a "dialect" abstraction together with tooling to define and transform custom dialects.

This MR makes two changes:
- It adds MLIR as an optional dependency to `cmake`. It's disabled by default. The flag `-DGQE_ENABLE_QUERY_COMPILER=ON` enables the dependency.
- It includes three integration tests for MLIR: a CPU Engine test, a GPU Engine test, and a GPU Standalone test.

The integration tests showcase three different approaches to compilation with MLIR, and include references to MLIR source code in the comments.

The MR is a continuation of the query compiler MRs (https://gitlab-master.nvidia.com/Devtech-Compute/gqe/-/merge_requests/191).

## Description of Integration Tests

The CPU Engine test generates a `main` function that calls `printf`. It compiles this code with LLVM. The code is run with the MLIR `ExecutionEngine`.

The GPU Engine test generates a `main` host function, and a `test` device kernel. The `main` function launches the `test` kernel. The `test` kernel calls `printf`. All of this code is generated using MLIR. The test compiles this code with LLVM and NVVM. The code is run with the MLIR `ExecutionEngine`.

The GPU Standalone test generates a `test` device kernel that calls `printf`. The host `main` function is hand-written. The host `main` compiles `test` using NVVM. After that, `main` launches the `test` kernel by loading it from a `cubin` using our own code, i.e., without the MLIR `ExecutionEngine`. Note that these module loading functions are part of the CUDA driver API, and not available in the CUDA runtime API.