Update MLIR to version `20.1.2` (Devtech-Compute/gqe!249)
This MR updates to the most recent MLIR version, `20.1.2`. It fixes breakage caused by upstream API changes.

Note that the `mlir_gpu_engine` test remains broken. Upstream MLIR is missing the `mlir::vector::registerConvertVectorToLLVMInterface()` function. However, this function has been added in the `main` branch. As I expect it to be available in the next release and this test is not critical, we can fix it in a later MLIR update.

The Docker new container is pushed as `gitlab-master.nvidia.com:5005/devtech-compute/gqe:24.12-cuda12.6-mlir-20.1.2-x86_64`, i.e., with a temporary MLIR suffix.