Improve compute communication overlap when running with multiple workers (Devtech-Compute/gqe!207)
When increasing worker count we end up with bad overlap between memcpy scheduled by the in-memory read task and the dependent kernels which are executed after. This is because the scheduler on the GPU will interleave the memcopies from different workers, leading to a given worker waiting longer for the in memory read task to complete. Ideally, we want a given worker to finish its memcopies as early as possible so that the data dependent kernel on that worker can run.

[Perf numbers on GH200](https://docs.google.com/spreadsheets/d/1gVqMPtfXfrflioQllIW9tON_aRvHLM5r38LcrvQSS7g/edit?usp=sharing)