Masked/Vectorized groupby (Devtech-Compute/gqe!227)
Motivation:

TPC-H Q1 has a non-selective filter followed by an aggregation. When we apply the filter and gather the result, we read and write almost all the rows. This is quite wasteful, especially because this is on the large lineitem table.

Current solution:

We take an optional condition expression in the aggregate relation, and we apply the condition to get the `active_mask` and use it directly in the groupby kernel, without materializing the rows that pass the condition.

Limitations :

* Evaluate expression within the aggregate task cannot use active_mask (cudf limitation). This means, we evaluate expressions before applying the `active_mask` on all the rows.
* Pure reductions cannot use the `active_mask` (cudf limitation)
* If we merge this, we cannot use shared memory groupby available in cudf-24.12 upstream and have to maintain it ourselves.

Impact:

* Q1 get a 10% improvement.
* Q15, Q6 has a similar pattern of filter+aggregate. But, Q15 has highly selective filter. Q6 has a reduction.

Future work:

* It is more common to have filter followed by join, investigate whether this optimization would be beneficial there. The work would be more as cudf's join doesn't allow an active_mask, hence, could require a re-implementation of join.
* If yes, we could have a better design where we can have the filter task return a `table + active_mask`. This `active_mask` can then be used in further operators. This can also be a building block for late materialization.