Format expressions as JSON (Devtech-Compute/gqe!256)
This MR adds an `expression_visitor` which serializes an `expression` tree to JSON.

The generated JSON string is contained in a single line, without extraneous white space or
indentation. It can be pretty printed using `jq .`.

## Limitations

Only the unary_operator and binary_operator types contained in TPC-H partial
filters are explicitly supported, e.g., an `equal_expression` is represented as:

    {"expression_type":"and_expression", ...}

Unsupported unary_operator and binary_operator types are represented using the
casted int value of the enum.

## Tests

No tests because this is only meant to be used in debugging output.

## Notes

There already exists `expression::to_string`, however, the output is difficult to read for complex expressions. JSON can be easily pretty-printed with external tools.