Upgrade Substrait to version 0.66.1 (Devtech-Compute/gqe!261)
## Changes to the transformation of Substrait protobuf messages to logical query plans

- `substr` function renamed to `substring`
- `JOIN_TYPE_SINGLE`, `JOIN_TYPE_SEMI`, and `JOIN_TYPE_ANTI` no longer exist; replaced with corresponding left join
- `LIMIT` and `OFFSET` clauses support expressions
- Grouping expressions are declared as references

Note: Needs to be tested with new plans in `/home/gqe/tpch/substrait_plans/with_filter_predicate_pushdown_to_read` (This is also why the pipeline fails at the moment)

## Impact on performance

- Tested with SF100
- Overall difference is 0.4%
- Q4 is 1.13x slower
- Q6 is 1.33x faster 
- Details in https://sc.talos.nvidia.com/view/home/scratch.vrosenfeld_gpu/notes/2025-05-09/new_substrait_plans/new_substrait_plans.html

## LIMIT and OFFSET expressions

Previously the values supplied to `LIMIT` and `OFFSET` clauses in the SQL queries were represented as integers. This has now been deprecated. Instead, the values are represented as a Substrait expression.

These values are supplied to `gqe::fetch_relation` which at the moment also represents them as integers. To support expressions in a generic way, two options are possible:

- Fold them to constants while transforming the Substrait messages to a logical plan.
- Change `gqe::fetch_relation` to accept `gqe::expression` and then perform a constant folding pass for all expressions in the optimizer.

For now, the code just checks if the Substrait expression is a literal and throws an exception otherwise.

## Tests

Tested by query execution scripts.

Closes #136