Read schema from catalog when optimizing partial filter expressions (Devtech-Compute/gqe!267)
This MR fixes a bug, so that the partial filter of a read relation is applied to the base table schema of the read relation and not the projected columns during the optimization of a logical query plan.

Closes #131

This is a revised version of https://gitlab-master.nvidia.com/Devtech-Compute/gqe/-/merge_requests/253. Compared to https://gitlab-master.nvidia.com/Devtech-Compute/gqe/-/merge_requests/253, users do not have to supply the base table schema to `read_relation`. Instead, it is read directly from the catalog.

## Changes

- The actual fix: Update `optimization_rule::rewrite_relation_expressions` to evaluate the partial filter of a `read_relation` on the column types of the entire table, instead of the projected columns from the `read_relation`
- Make `optimization_rule::rewrite_relation_expressions` an instance method, so that it has access to the `optimization_rule::_catalog`
- Add method `catalog::column_types` to retrieve all column types of table
- Make `join_children_swap_helpers::swap_join_keys_inplace` a private method of `join_children_swap` and remove the helper class `join_children_swap_helpers`
- Update `rewrite_rule_test.cpp` and `string_to_int_literal_test.cpp` to register a test table with the catalog, so that it can be accessed in `optimization_rule::rewrite_relation_expressions`

## Tests

- Added specialized test in `string_to_int_literal_test.cpp`