Enhance uniqueness propagation for join (Devtech-Compute/gqe!265)
Updated optimizer rule with learnings from [gqe-python MR!60](https://gitlab-master.nvidia.com/haog/gqe-python/-/merge_requests/60) to enable more opportunities for propagation. Updated unit test.

Previously, column uniqueness properties are propagated from the input columns to the output columns of a join relation only if all input columns involved in equality comparisons in the join predicate are unique. However, @matal realized that a more permissive condition could be used, something like: `If one side is unique, then the other side columns get to keep their uniqueness`. For example, given table `A` with columns `a1` (unique) and `a2` (not unique), and table `B` with columns `b1` (unique), `b2` (not unique), a join of `A` and `B` with predicate `A.a1 = B.b2` cannot result in the duplication of any rows in `B`'s columns since `A.a1` is unique. Hence, the output column corresponding to `B.b1` is still unique. On the other hand, since `B.b2` is not unique, rows in `A`'s columns can be duplicated in the output so the uniqueness of `A.a1` does not propagate.

SF100:

* All query results validated
* Q3, Q5, and Q10 saw improvement in end-to-end flow (41%, 57%, and 15%, respectively). However, e2e flow only matches hand-coded plan performance for Q3 and Q5, while hand-coded plan is still faster for Q10.
* No significant change for other queries

SF1k:

* Q3 end-to-end flow improved 53%