Add support for numa_pinned_memory (Devtech-Compute/gqe!268)
This MR adds support for NUMA pinned memory. 

The main difference between NUMA memory and this is that we register the NUMA binded memory with CUDA. 

These changes are targeted towards testing on GB200, where we have two Grace CPUs, and GQE would use the memory of both if we don't specify NUMA memory. But we would lose the option of pinned memory, and this option resolves that.