Fix ownership of RAII lock (Devtech-Compute/gqe!292)
A unique lock for the copy engine mutex was previously made a class member in `in_memory_read_task` . During multi-worker execution, when this lock acquires the mutex for the copy engine and the thread encounters an exception, the destructor of the task object is not called as the task is owned by the main thread. This can then lead to a deadlock as the copy engine's mutex is not released.

This change limits the lifetime of the unique lock to within `in_memory_read_task::execute` so that when an exception is thrown and the stack is unwound, the lock object goes out of scope and releases the mutex.

This fixes [Deadlock with use_overlap_mtx and multiple workers (#164) · Issues · Devtech-Compute / GQE · GitLab](https://gitlab-master.nvidia.com/Devtech-Compute/gqe/-/issues/164).

Also including the fix for a leaking `cudaEvent_t` handle under the same situation.

As a bonus, this change removes 3 private methods and 2 private attributes from the interface of `in_memory_read_task`.