CMAKE fixes (Devtech-Compute/gqe!297)
This MR addresses two issues with the current CMAKE:

- The CMAKE_CUDA_ARCHITECTURES was getting reset by RAPIDS cmake. When one sets its value using `set(CMAKE_CUDA_ARCHITECTURES "")`, it is only set locally within the local cmake, and its value is not carried to other included libraries cmake unless you cache it.
Currently, we weren't caching it. Consequently, RAPIDS CMAKE wasn't seeing it, and resetting it. 

This was not an issue when we explicitly define CMAKE_CUDA_ARCHITECTURES, or if we install gqe via gqe-python. But for some reason, installing gqe stand-alone did not cache it.
- groupby_test did not have the right include path, which was causing a compile error.