Perfect aggregation (Devtech-Compute/gqe!290)
This improves performance over `cudf::gropuby`.  As a next step, we need to also use per-thread shmem in the scatter_aggregate (the idea of @zhengruw).

ref: https://gitlab-master.nvidia.com/Devtech-Compute/tpch-cuda/-/merge_requests/169

Run on GH200 with SF1k, including only queries for which perfect hashing in aggregation is possible, here are the results:

| Query | perfect hashing | main         | improvement |
|-------|-----------------|--------------|-------------|
| 1     | 1.244039771     | 1.360553149  | 9.37%       |
| 2     | 0.1157297729    | 0.1116729263 | -3.51%      |
| 3     | 0.7884616455    | 0.743748291  | -5.67%      |
| 6     | 0.3670389587    | 0.3616741943 | -1.46%      |
| 15    | 0.2659441772    | 0.2818131653 | 5.97%       |
| 17    | 0.2497293091    | 0.2266714294 | -9.23%      |
| 19    | 1.163070898     | 1.139555566  | -2.02%      |
| 20    | 0.4797779358    | 0.4754730713 | -0.90%      |

We see that generally perfect hashing is not an end-to-end improvement on the query performance.  This should improve after the "per-thread private shmem" feature is added from @zhengruw.