Feat: add main branch CI pipelines. (Devtech-Compute/gqe!315)
This MR is intended to create a CI pipeline for main branch merge requests.

The CI gitlab runner is a temporary workaround plan to replace NDash as NDash is EOL. The new CI is running on a GH200 node from Colossus cluster. The expiration date of the Colossus machine is 10/8. For detailed setup instructions for a gitlab runner running on Colossus, please refer [here](https://confluence.nvidia.com/display/DevtechCompute/Colossus+Gitlab+CI+Setup).

## Changes

* Add `.gitlab-ci.yml` for CI pipeline.
* The CI pipeline would run unit tests for debug and release mode. It would also run tpch end-to-end tests for SF1. SF100 and SF1K could be extended later for the CI.

## Usage

* `/tpch_scratch` has the following directory: `datasets`, `reference_results` , `substrait_plans` with the same directory hierarchy as nfs `/home/tpc/tpch` . But only SF1 and SF100 datasets are copied here.
* If more CI jobs are needed, just reference the `/tpch_scratch` to access necessary files.