Use branch specific Dockerfile in CI (Devtech-Compute/gqe!342)
## Goal and reason

We cannot test changes to Dockerfile and conda environment in CI under the current setup without modifying gitlab pipelines.

## Problem statement

CI should allow team to check branch specific changes to the environment before merge without having to modify the gitlab pipeline files.

## Solution approach

The solution checks if the CI is triggered from a MR and includes changes to the `Dockerfile` or conda environment files. If this is the case it tries to run CI on a specific docker image with tag `CI_COMMIT_REF_SLUG` . See https://docs.gitlab.com/ci/variables/predefined_variables/ on how to name this image. If this is not the case then we pull the `latest_x86_64` image.

The tag should be named using the following rule: Branch name in lowercase, shortened to 63 bytes, and with everything except `0-9` and `a-z` replaced with `-`. No leading / trailing `-`.

## Alternative approaches

_N/A_

## Impact (feature / bug / performance)

This is a infra feature. See https://gitlab-master.nvidia.com/Devtech-Compute/gqe/-/merge_requests/309#note_41150919 for a concrete example.

## Issues closed

_N/A_

## Tests

MR is tested using gitlab pipelines.

## References

* https://docs.gitlab.com/ci/variables/predefined_variables/
* https://docs.gitlab.com/ci/jobs/job_rules/

## Review checklist

(to be signed off by the reviewer)

- [x] Assigned self as reviewer in GitLab sidebar.
- [ ] MR branch is in the `DevTech-Compute/gqe` repository (i.e., not a fork), and is prefixed with `$user-`.
  - Exception: Author has not been granted write permissions.
- [x] MR title is ["syntactically correct"](https://cbea.ms/git-commit). E.g., "Optimize the aggregate operator for high group cardinality".
  - Less than 50 chars.
  - Capitalized.
  - Imperative mood.
  - Descriptive of the change.
- [x] MR goal, problem, and approach make sense to me (as described above).
- [x] MR is self-contained. I.e., doesn't rely on follow-up MRs to be minimially functional.
- [x] MR solves _one_ (sub-)issue. Big issue is split into multiple self-contained MRs.
- [x] All files include [the NVIDIA copyright header](https://confluence.nvidia.com/display/LEG/NVIDIA+Proprietary) and year is up-to-date.
- [x] Code style adheres to the [C++ Core Guidelines](https://isocpp.github.io/CppCoreGuidelines/CppCoreGuidelines) and [libcudf developer guidelines](https://github.com/rapidsai/cudf/blob/main/cpp/doxygen/developer_guide/DEVELOPER_GUIDE.md).
  - MLIR code instead adheres to the [LLVM coding standards](https://llvm.org/docs/CodingStandards.html).
- [x] Code is DRY. I.e., uses existing GQE / cuDF / MLIR infrastructure and utilities.
- [x] Code is documented.
  - [Doxygen API comments](https://github.com/rapidsai/cudf/blob/main/cpp/doxygen/developer_guide/DOCUMENTATION.md).
  - Algorithm / data structure description (if relevant).
- [x] Code includes CI unit and/or integration tests (if relevant).
- [x] MR is tested (see section Tests) and doesn't break the gqe-python main branch.
- [x] Stale MR branches are cleaned up. E.g., `this-MR-branch-v1`.