Adds sharing of in-memory table between processes for multi-GPU (Devtech-Compute/gqe!309)
## Goal and reason

This MR adds sharing of in-memory table between processes for multi-process execution.

## Solution approach

This is implemented by using shared CPU memory with the help of the library [Boost](https://www.boost.org/doc/libs/1_88_0/doc/html/interprocess.html), that provides functionality to safely share C++ data structures and perform address translation across process.

The root rank reads all the columns, and writes to shared CPU memory. The columns can be compressed/uncompressed. While, the other ranks store pointers to these columns. We also store zone-maps in the shared CPU memory, which are copied to the device during the subsequent read tasks.


## Alternative approaches

The alternatives considered were using MPI shared memory, or directly allocating using `mmap`. But, these required us to manually handle complex object creation and pointer translation across processes. Hence, Boost library was used to provide us this functionality.

## Impact (feature / bug / performance)

This is a bring-up MR, and we are tracking some design/bugs/cleanups in the [issue](https://gitlab-master.nvidia.com/Devtech-Compute/gqe/-/issues/196), which will be addressed in subsequent MRs. 

This enabled us to showcase multi-GPU scaling performance of 1.5x when going from one to two ranks.

## Issues closed

#140 

## Tests

This was tested on SF1k. Some parameter configurations fail, and are being tracked in #194 and #175 

## Review checklist

(to be signed off by the reviewer)

 - [x] Assigned self as reviewer in GitLab sidebar.
 - [ ] MR branch is in the `DevTech-Compute/gqe` repository (i.e., not a fork),
       and is prefixed with `$user-`.
   - Exception: Author has not been granted write permissions.
 - [x] MR title is ["syntactically correct"](https://cbea.ms/git-commit). E.g.,
       "Optimize the aggregate operator for high group cardinality".
   - Less than 50 chars.
   - Capitalized.
   - Imperative mood.
   - Descriptive of the change.
 - [x] MR goal, problem, and approach make sense to me (as described above).
 - [x] MR is self-contained. I.e., doesn't rely on follow-up MRs to be
       minimially functional.
 - [x] MR solves *one* (sub-)issue. Big issue is split into multiple
       self-contained MRs.
 - [x] All files include [the NVIDIA copyright header](https://confluence.nvidia.com/display/LEG/NVIDIA+Proprietary) and year is up-to-date.
 - [x] Code style adheres to the [C++ Core Guidelines](https://isocpp.github.io/CppCoreGuidelines/CppCoreGuidelines) and [libcudf developer guidelines](https://github.com/rapidsai/cudf/blob/main/cpp/doxygen/developer_guide/DEVELOPER_GUIDE.md).
   - MLIR code instead adheres to the [LLVM coding standards](https://llvm.org/docs/CodingStandards.html).
 - [ ] Code is DRY. I.e., uses existing GQE / cuDF / MLIR infrastructure and
       utilities.
 - [x] Code is documented.
   - [Doxygen API comments](https://github.com/rapidsai/cudf/blob/main/cpp/doxygen/developer_guide/DOCUMENTATION.md).
   - Algorithm / data structure description (if relevant).
 - [x] Code includes CI unit and/or integration tests (if relevant).
 - [x] MR is tested (see section Tests) and doesn't break the gqe-python main
       branch.
 - [x] Stale MR branches are cleaned up. E.g., `this-MR-branch-v1`.