Update cudf to final commit of 25.10 compatible with nvcomp 5.0 (Devtech-Compute/gqe!336)
## Goal and reason

This MR updates cudf to a version that is compatible with nvcomp 5.0

## Problem statement

We want to move to a custom version of nvcomp in GQE (that is based on nvcomp 5.0). But 5.0 introduces breaking API changes. Due to the dependency chain, we also want to ensure cudf is compatible with nvcomp 5.0.

## Solution approach

The solution is to first update cudf to a version that is compatible with nvcomp 5.0, and then update to the custom nvcomp version.

But cudf now requires GCC v14, which would require Ubuntu v24.04 as gcc v14 is not hosted on v22.04 apt repos.

There have also been other updates to the Conda environment to make it in line with cudf's build environment

In addition, this MR also makes changes to GQE to make it compatible with the breaking changes in cudf, including:

- Replaced calls to `left_semi_join` and `left_anti_join` that are deprecated. Functionailty is now implemented in `cudf::filtered_join` object.
- `convert_fixed_to_floating` was moved from `unary.hpp` to `conv.hpp`. Modified include files to account for this.



## Alternative approaches

N/A

## Impact (feature / bug / performance)

This MR by itself would significantly improve compression performance on the base nvcomp. At present, compression performance is very bad in batched mode if we test with the nvcomp version currently included in Conda. Performance differences are shown below:
![image](/uploads/2e99bf1934c15df3f2fa964dd90fd722/image.png)

## Issues closed

None

## Tests

Tested with SF100 on parameter sweep.

## Review checklist

(to be signed off by the reviewer)

 - [x] Assigned self as reviewer in GitLab sidebar.
 - [ ] MR branch is in the `DevTech-Compute/gqe` repository (i.e., not a fork),
       and is prefixed with `$user-`.
   - Exception: Author has not been granted write permissions.
 - [x] MR title is ["syntactically correct"](https://cbea.ms/git-commit). E.g.,
       "Optimize the aggregate operator for high group cardinality".
   - Less than 50 chars.
   - Capitalized.
   - Imperative mood.
   - Descriptive of the change.
 - [x] MR goal, problem, and approach make sense to me (as described above).
 - [x] MR is self-contained. I.e., doesn't rely on follow-up MRs to be
       minimially functional.
 - [x] MR solves *one* (sub-)issue. Big issue is split into multiple
       self-contained MRs.
 - [x] All files include [the NVIDIA copyright header](https://confluence.nvidia.com/display/LEG/NVIDIA+Proprietary) and year is up-to-date.
 - [x] Code style adheres to the [C++ Core Guidelines](https://isocpp.github.io/CppCoreGuidelines/CppCoreGuidelines) and [libcudf developer guidelines](https://github.com/rapidsai/cudf/blob/main/cpp/doxygen/developer_guide/DEVELOPER_GUIDE.md).
   - MLIR code instead adheres to the [LLVM coding standards](https://llvm.org/docs/CodingStandards.html).
 - [x] Code is DRY. I.e., uses existing GQE / cuDF / MLIR infrastructure and
       utilities.
 - [x] Code is documented.
   - [Doxygen API comments](https://github.com/rapidsai/cudf/blob/main/cpp/doxygen/developer_guide/DOCUMENTATION.md).
   - Algorithm / data structure description (if relevant).
 - [x] Code includes CI unit and/or integration tests (if relevant).
 - [x] MR is tested (see section Tests) and doesn't break the gqe-python main
       branch.
 - [x] Stale MR branches are cleaned up. E.g., `this-MR-branch-v1`.